requirements.txt
absl-py==2.1.0
accelerate==1.4.0
addict==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
airportsdata==20250224
albucore==0.0.24
albumentations==2.0.6
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
basicsr==1.4.2
beautifulsoup4==4.13.4
bitsandbytes==0.45.2
blake3==1.0.4
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
clip-anytorch==2.6.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
compressed-tensors==0.9.1
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cupy-cuda12x==13.4.1
cycler==0.12.1
Cython==3.0.12
datasets==3.4.1
dctorch==0.1.2
debugpy==1.8.13
decorator==5.2.1
depyf==0.18.0
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email_validator==2.2.0
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.14.0
fire==0.7.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h11==0.14.0
hf_transfer==0.1.9
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.1.0
interegular==0.3.3
ipykernel==6.29.5
ipympl==0.9.7
ipython==8.34.0
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
jiwer==3.1.0
jmespath==0.10.0
jsonmerge==1.9.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
k-diffusion==0.1.1.post1
kiwisolver==1.4.8
kornia==0.7.0
lark==1.2.2
lazy_loader==0.4
lightning-utilities==0.12.0
llvmlite==0.43.0
lm-format-enforcer==0.10.11
lmdb==1.6.2
lpips==0.1.4
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.3
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
open-clip-torch==2.24.0
openai==1.68.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
pluggy==1.5.0
prettytable==3.16.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pybind11==2.13.6
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.1.2
pytz==2023.4
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.13.0
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.4.2
rich-toolkit==0.13.2
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.24.1
setproctitle==1.3.5
shapely==2.1.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.45.3
stringzilla==3.12.5
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.20.0a20250326
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminaltables==3.1.10
tifffile==2025.3.13
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
tomli==2.2.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchdiffeq==0.2.5
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.52.2
triton==3.1.0
trl==0.15.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.7.3
wandb==0.19.8
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
